+++ Assets/Scripts/Audio Scripts/EnvMusic.cs
@@ -1,0 +1,71 @@
+using Unity.VisualScripting.Antlr3.Runtime;
+using UnityEngine;
+
+public class EnvMusic : MonoBehaviour
+{
+    public AudioClip fallbackMusic; // 
+    public AudioClip GrassRoomMusic; //
+
+    // Volume multipliers for each surface type (adjust these values based on the loudness of your recordings)
+    public float fallbackVolume = 0.2f; // Volume for fallback music
+    public float grassRoomVolume = 0.2f;  // Grass music volume (default)
+    public Transform listenerTransform;    // Reference to the camera that has the listener
+    private CharacterController characterController;
+    private float volumeMultiplier = 0.03f; // Multiply all footstep sounds by this value to bring it to a background level.
+    private AudioSource audioSource;
+    AudioClip selectedMusic;
+
+    private void Start()
+    {
+        audioSource = listenerTransform.GetComponent<AudioSource>();
+        characterController = GetComponent<CharacterController>();
+    }
+
+    private void Update()
+    {
+
+    }
+
+    private void OnTriggerEnter(Collider other)
+    {
+        selectedMusic = findCorrectMusicTrack(other);
+
+        if (selectedMusic != null && !audioSource.isPlaying)
+        { // It won't play if an audio is not found, or it won't play if a music is already playing.
+            audioSource.clip = selectedMusic;
+            audioSource.Play();
+        }
+
+    }
+    private void OnTriggerExit(Collider other)
+    {
+        selectedMusic = findCorrectMusicTrack(other);
+
+        if (audioSource.clip == selectedMusic)
+        { // It won't play if an audio is not found, or it won't play if a music is already playing.
+            audioSource.clip = selectedMusic;
+            audioSource.Stop();
+        }
+
+    }
+
+    private AudioClip findCorrectMusicTrack(Collider other)
+    {
+        AudioClip selectedTrack = null;
+
+        if (other.CompareTag("MusicGrassroom"))
+        {
+            selectedTrack = GrassRoomMusic;
+        }
+
+        /*else if (other.CompareTag("MusicStoneroom")){
+            selectedMusic = StoneRoomMusic;
+            audioSource.clip = selectedMusic; 
+            selectedMusic.Play();
+        }*/
+
+        return selectedTrack;
+    }
+
+
+}
