--- Assets/Scripts/Audio.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bf41e51e307393145af430c3b8f51e2c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
