--- Assets/Scripts/Audio/EnvMusic.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-
-public class EnvMusic : MonoBehaviour
-{
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
