--- Assets/Scripts/Audio/EnvMusic.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ae42bc5a1d7d85b4ba90e9b867444092
