--- Assets/Scripts/Audio/FootstepSound.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 19095c33f169b4e4ea9347275102b631
