+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Grass
   - Stone
   - Dirt
+  - MusicGrassroom
   layers:
   - Default
   - TransparentFX
